requirements.txt
# Core Audio Processing
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0

# Wake Word Detection
pvporcupine>=3.0.0

# Speech Recognition
# whisper.cpp will be handled separately via bindings
faster-whisper>=1.0.0  # Fast local STT with quantization
openai>=1.3.0  # For Realtime API
websockets>=12.0

# Voice Activity Detection
silero-vad>=4.0.0  # Speech start/stop detection

# Natural Language Understanding
spacy>=3.7.0
# Download model with: python -m spacy download en_core_web_sm

# Vector Database / Memory
chromadb>=0.4.18
# faiss-cpu>=1.7.4  # Alternative to ChromaDB

# Task Scheduling
APScheduler>=3.10.4

# Calendar Integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Text-to-Speech
piper-tts>=1.2.0
edge-tts>=6.1.9
# coqui-tts>=0.22.0  # Alternative TTS

# Desktop Notifications (Windows)
windows-toasts>=1.1.0; platform_system=="Windows"
# desktop-notifier>=3.5.0  # Cross-platform alternative

# Web Automation
playwright>=1.40.0

# Desktop UI
PySide6>=6.6.0
PySide6-Addons>=6.6.0

# Utilities
PyYAML>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.5.0
loguru>=0.7.2
psutil>=5.9.6  # System and process utilities

# Keyboard Hotkeys
keyboard>=0.13.5

# Audio Format Handling
soundfile>=0.12.1
pyaudio>=0.2.14

# Development Tools
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Build Tools for C++ Extension
cmake>=3.27.0
pybind11>=2.11.1

# Windows Audio Control (Fallback)
pycaw>=20230407; platform_system=="Windows"

# Audio Playback
playsound==1.2.2  # Simple MP3/WAV playback

# Configuration
PyYAML>=6.0.1  # YAML parser for config files

